refactor: Extract NetLineBody and NetLine classes from Ssta
- Create src/net_line.hpp with NetLineBody and NetLine classes
- Move NetLineIns type alias to net_line.hpp
- Remove NetLineBody and NetLine definitions from Ssta class header
- Update Ssta class to use NetLineIns instead of Ins
- Improve separation of concerns and code organization

This refactoring:
- Reduces Ssta class header size (81 lines removed)
- Improves single responsibility principle compliance
- Makes NetLine classes reusable and independently testable
- Maintains all existing functionality

Fixes #98
