+++ include/nhssta/ssta.hpp
@@ -16,6 +16,7 @@
 #include <nhssta/ssta_results.hpp>
 
 #include "../src/gate.hpp"
+#include "../src/net_line.hpp"
 
 // Forward declaration for Parser (internal implementation detail)
 // Parser is defined in global namespace in src/parser.hpp
@@ -36,84 +37,6 @@ class Ssta {
     void read_bench();
 
    private:
-    typedef std::vector<std::string> Ins;
-
-    class NetLineBody {
-       public:
-        NetLineBody() = default;
-        virtual ~NetLineBody() = default;
-
-        void set_out(const std::string& out) {
-            out_ = out;
-        }
-        [[nodiscard]] const std::string& out() const {
-            return out_;
-        }
-
-        void set_gate(const std::string& gate) {
-            gate_ = gate;
-        }
-        [[nodiscard]] const std::string& gate() const {
-            return gate_;
-        }
-
-        [[nodiscard]] const Ins& ins() const {
-            return ins_;
-        }
-        Ins& ins() {
-            return ins_;
-        }
-
-       private:
-        std::string out_;
-        std::string gate_;
-        Ins ins_;
-    };
-
-    class NetLine {
-       public:
-        NetLine()
-            : body_(std::make_shared<NetLineBody>()) {}
-        explicit NetLine(std::shared_ptr<NetLineBody> body)
-            : body_(std::move(body)) {
-            if (!body_) {
-                throw RuntimeException("NetLine: null body");
-            }
-        }
-
-        NetLine(const NetLine&) = default;
-        NetLine(NetLine&&) noexcept = default;
-        NetLine& operator=(const NetLine&) = default;
-        NetLine& operator=(NetLine&&) noexcept = default;
-        ~NetLine() = default;
-
-        NetLineBody* operator->() const {
-            return body_.get();
-        }
-        NetLineBody& operator*() const {
-            return *body_;
-        }
-
-        bool operator==(const NetLine& rhs) const {
-            return body_.get() == rhs.body_.get();
-        }
-        bool operator!=(const NetLine& rhs) const {
-            return !(*this == rhs);
-        }
-        bool operator<(const NetLine& rhs) const {
-            return body_.get() < rhs.body_.get();
-        }
-        bool operator>(const NetLine& rhs) const {
-            return body_.get() > rhs.body_.get();
-        }
-
-        [[nodiscard]] std::shared_ptr<NetLineBody> get() const {
-            return body_;
-        }
-
-       private:
-        std::shared_ptr<NetLineBody> body_;
-    };
 
     void read_dlib_line(Parser& parser);
     void read_bench_input(Parser& parser);
@@ -122,7 +45,7 @@ class Ssta {
     void set_dff_out(const std::string& out_signal_name);
     void connect_instances();
     [[nodiscard]] bool is_line_ready(const NetLine& line) const;
-    void set_instance_input(const Instance& inst, const Ins& ins);
+    void set_instance_input(const Instance& inst, const NetLineIns& ins);
     void check_signal(const std::string& signal_name) const;
 
     void node_error(const std::string& head, const std::string& signal_name) const;
